update/remove some rails legacy tests
