+++ spec/integration/rails/graphql/relay/range_add_spec.rb
@@ -82,17 +82,8 @@ describe GraphQL::Relay::RangeAdd do
     Class.new(GraphQL::Schema) do
       self.query(query)
       self.mutation(mutation)
-      if TESTING_INTERPRETER
-        self.use(GraphQL::Analysis::AST)
-        self.use(GraphQL::Execution::Interpreter)
-      end
-      self.cursor_encoder(PassThroughEncoder)
 
-      if TESTING_INTERPRETER
-        use GraphQL::Execution::Interpreter
-        use GraphQL::Analysis::AST
-        use GraphQL::Pagination::Connections
-      end
+      self.cursor_encoder(PassThroughEncoder)
     end
   }
